FBNBenchmark {FBNNet}	R Documentation

A benchmark type function to test a complete process of FBN model

Description

A benchmark type function to test a complete process of FBN model

Usage

FBNBenchmark(booleanNetwork, shorttrainingtimeSeries, shorttestseries,
  originaltestseries, interval = 1, temporal = 1, convertToFBN = TRUE,
  existingFBNCube = NULL)

Arguments

booleanNetwork
A Boolean Network object

shorttrainingtimeSeries
A training timeseries matrix data set

shorttestseries
A test timeseries matrix data set

originaltestseries
A original timeseries matrix data set

interval
The internal gat between time points

temporal
A temporal setting with default =1

convertToFBN
if it is a need to convert the boolean network to FBN Network. If the boolean network object is not a FBNNetwork object, then it must be set to TRUE

existingFBNCube
An existing FBN cube object

decay
The decay time step for protein decay

Value

A list of object

Examples

coming later

[Package FBNNet version 0.0.0.9000 Index]