Functions related to Bioinformatics

A collection of functions that used to handle bioinformatics related tasks

Arguments

files

The file folder path or directory that contains the affy files (*.cel)

useGCRMA

optional, if true it use GCRMA method to normalize the affy data, otherwise, use the default RMA method. GCRMA, A bias-corrected RMA; FALSE, use RMA, which is based on Robust Multi-Chip' average

cdfname

The name of CDF that is associated with the affy data.

cellDirectory

A directory that contains affy raw data

outPutEvalue

optional to write the result into disk

rawData

the raw data that is the output from the method getRelatedAffyRawData

ges_assess_no

the GSE identity id, such as GSE35635

References

Chen et al.(2018), Front. Physiol., 25 September 2018, (Front. Physiol.)

Bioconductor (2019), https://www.bioconductor.org/

Author

Leshi Chen, leshi, chen@lincolnuni.ac.nz, chenleshi@hotmail.com

Examples

exprs <- downloadGenomeDataExprs(GSE35635)
#> Error in sprintf("Enter downloadGenomeDataExprs zone: ges_assess_no=%s",     ges_assess_no): object 'GSE35635' not found
exprs
#> Error in eval(expr, envir, enclos): object 'exprs' not found