FBNNet-package {FBNNet}	R Documentation

FBNNet is a R package for analyzing time series data under Fundamental Boolean Network Models

Description

What the package does (one paragraph).

Details

The DESCRIPTION file: This package was not yet installed at build time.

Index: This package was not yet installed at build time.
~~ An overview of how to use the package, including the most important functions ~~

Author(s)

NA

Maintainer: NA

References

~~ Literature or other references for background information ~~

See Also

~~ Optional links to other man pages, e.g. ~~ ~~ <pkg> ~~

Examples

~~ simple examples of the most important functions ~~

[Package FBNNet version 0.0.0.9000 Index]