FBNNetwork.Graph {FBNNet}	R Documentation

Executing a list of processes in parallel with decreasing list

Description

Executing a list of processes in parallel with decreasing list

Usage

FBNNetwork.Graph(fbnNetwork, type = "static", timeseriesMatrix = NULL,
  fromTimePoint = 1, toTimePoint = 5, networkobject = NULL)

Arguments

fbnNetwork
A FBN Network object

type
A type of object and its value is type=static | dynamic | staticSlice

timeseriesMatrix
A sample timeseries matrix

fromTimePoint
The time point at the beginning

toTimePoint
The time point at the end

Value

No return

Examples

coming later

[Package FBNNet version 0.0.0.9000 Index]