A method to draw attractors in the form of dynamic Network

A method to draw attractors in the form of dynamic Network

FBNNetwork.Graph.DrawAttractor(fbnNetwork, FBMAttractors, index = 1)

Arguments

fbnNetwork

An object of FBNNetwork

FBMAttractors

A list of attractors extracted via searchForAttractors.

index

The index of an attractor in the list FBMAttractors.

Examples

require(BoolNet)
data('ExampleNetwork')
initialStates<-generateAllCombinationBinary(ExampleNetwork$genes)
trainingseries<-genereateBoolNetTimeseries(ExampleNetwork,
                                           initialStates,
                                           43,
                                           type='synchronous')
cube<-constructFBNCube(target_genes = ExampleNetwork$genes,
                       conditional_genes = ExampleNetwork$genes,
                       timeseriesCube = trainingseries,
                       maxK = 4,
                       temporal = 1,
                       useParallel = FALSE)
#> INFO [2022-04-08 11:05:09] Enter constructFBNCube zone: 
#>         target_genes=5 genes and they are Gene1, Gene2, Gene3, Gene4, Gene5,
#>         conditional_genes=5 genes and they are Gene1, Gene2, Gene3, Gene4, Gene5,
#>         data_length=32,
#>         maxK=4, 
#>         temporal=1,
#>         useParallel=FALSE
#> INFO [2022-04-08 11:05:09] Leave constructFBNCube zone.
NETWORK2<-mineFBNNetwork(cube,ExampleNetwork$genes)
#> INFO [2022-04-08 11:05:09] Enter mineFBNNetwork zone: genes=Gene1, Gene2, Gene3, Gene4, Gene5, useParallel=FALSE
#> INFO [2022-04-08 11:05:09] Enter search_FBN_core zone: useParallel=FALSE
#> INFO [2022-04-08 11:05:09] Leave search_FBN_core zone
#> INFO [2022-04-08 11:05:09] Enter mineFBNNetworkWithCores zone
#> INFO [2022-04-08 11:05:09] Enter mineFBNNetworkStage2 zone
#> INFO [2022-04-08 11:05:09] Leave mineFBNNetworkStage2 zone
#> INFO [2022-04-08 11:05:09] Enter convertMinedResultToFBNNetwork zone
#> INFO [2022-04-08 11:05:09] Leave convertMinedResultToFBNNetwork zone
#> INFO [2022-04-08 11:05:09] Leave mineFBNNetworkWithCores zone
#> INFO [2022-04-08 11:05:09] Leave mineFBNNetwork zone
attractor<-searchForAttractors(NETWORK2,
                               initialStates,
                               ExampleNetwork$genes)
print(attractor)
#> Discovered Attractors via Fundamental Boolean Model :
#> Genes are encoded in the following order::
#> Gene1 Gene2 Gene3 Gene4 Gene5:
#> 
#> Attractor 1 is a simple attractor consisting of 1 state(s):
#> 
#> | --< - - - - - |
#> v               ^
#> 0 0 0 0 1       |
#> |               |
#> v               ^
#> | - - - - - >-- |
#> 
#> 
#> Attractor 2 is a complex attractor consisting of 4 state(s):
#> 
#> | --< - - - - - |
#> v               ^
#> 1 0 0 0 0       |
#> |               |
#> 1 0 0 0 1       |
#> |               |
#> 1 1 0 0 1       |
#> |               |
#> 1 1 0 0 0       |
#> |               |
#> v               ^
#> | - - - - - >-- |
#> 
#> 
#> Attractor 3 is a simple attractor consisting of 1 state(s):
#> 
#> | --< - - - - - |
#> v               ^
#> 0 0 1 1 1       |
#> |               |
#> v               ^
#> | - - - - - >-- |
#> 
#> 
#> Attractor 4 is a complex attractor consisting of 5 state(s):
#> 
#> | --< - - - - - |
#> v               ^
#> 1 0 1 1 1       |
#> |               |
#> 1 0 1 0 1       |
#> |               |
#> 1 1 1 0 1       |
#> |               |
#> 1 1 1 0 0       |
#> |               |
#> 1 0 1 1 0       |
#> |               |
#> v               ^
#> | - - - - - >-- |
#> 
#> 
FBNNetwork.Graph.DrawAttractor(NETWORK2,attractor,2)