buildProbabilityTreeOnTargetGene {FBNNet}	R Documentation

Find all genes that are related to the target gene

Description

Find all genes that are related to the target gene

Usage

buildProbabilityTreeOnTargetGene(targetGene, currentState, previousState,
  timeseries, genes, matchedgenes, matchedexpression, maxK = 5,
  temporal = 1, allRelatedTargetGeneMeasures,
  allRelatedConditionalGenesMeasures, useParallel = FALSE)

Arguments

targetGene
The target gene

currentState
current combined gene state i.e. the gene state at t

previousState
previous combined gene state i.e. the gene state at t-1

genes
a list of genes that the related genes re mined from

matchedgenes
a list of pre matched genes

matchedexpression
previous matched expression string

maxK
The maximum depth the cube can go to. The default value is 5

temporal
1: d means temporal = t-1 ... t-d

allRelatedTargetGeneMeasures
all measures that are related to the target genes.

allRelatedConditionalGenesMeasures
all measures that are related to the conditional genes

timeseriesCube
a list of samples in which a sample contains gene and time points

An
option to turn on parallel

Value

A list of measures

Examples

mat1<-matrix(c("1","0","0","1","0","0","0","1","1"),3,3, dimnames=list(c("gene1","gene2","gene3"),c("1","2","3")))
mat2<-matrix(c("1","1","0","1","0","1","1","1","0"),3,3, dimnames=list(c("gene1","gene2","gene3"),c("1","2","3")))
listtest<-list(mat1,mat2)
completeState<-extractGeneStateFromTimeSeriesCube(listtest)
getCurrentState<-completeState
getpreviousState<-completeState
getCurrentState<-getCurrentState[,-1]#remove the first one
getpreviousState<-getpreviousState[,-length(colnames(getpreviousState))] #remove the last one
clusteringOnTargetGene("gene1",getCurrentState,getpreviousState,listtest,c("gene1","gene2","gene3"),1,FALSE)

[Package FBNNet version 0.0.0.9000 Index]