constructFBNCube {FBNNet}	R Documentation

Create an Orchard cube

Description

Create an Orchard cube

Usage

constructFBNCube(targetgenes, allgenes, timeseriesCube, maxK = 5,
  temporal = 1, useParallel = FALSE)

Arguments

targetgenes
A vector of genes that will be treated as target genes

allgenes
All genes that are available for building up the cube

timeseriesCube
A list of samples in which a sample is a matrix that contains gene states where genes in rows and time points in columms

maxK
The maximum level the cube can dig in

temporal
A value that used to be 1 indicates the previous steps the current one can depend on

useParallel
If it is TRUE, the constructing will run it in parallel, otherwise in a singl thread

Value

A Orcahrd cube that contains all precomputed measures

Examples

mat1<-matrix(c("1","0","0","1","0","0","0","1","1"),3,3, dimnames=list(c("gene1","gene2","gene3"),c("1","2","3")))
mat2<-matrix(c("1","1","0","1","0","1","1","1","0"),3,3, dimnames=list(c("gene1","gene2","gene3"),c("1","2","3")))
listtest<-list(mat1,mat2)
constructFBNCube(c("gene1","gene2"),c("gene1","gene2","gene3"),listtest,4,1,FALSE)

[Package FBNNet version 0.0.0.9000 Index]