convertToFBNNetwork {FBNNet}	R Documentation

Convert a boolean network object to fundamental boolean function object

Description

Convert a boolean network object to fundamental boolean function object

Usage

convertToFBNNetwork(network)

Arguments

network
A boolean network object

Value

An object of FBN network

Examples

coming later

[Package FBNNet version 0.0.0.9000 Index]