doNonParallelWorkDecrease {FBNNet}	R Documentation

Executing a list of processes in parallel with decreasing list

Description

Executing a list of processes in parallel with decreasing list

Usage

doNonParallelWorkDecrease(parallelFuc, listitems, unprocessedListitems, ...)

Arguments

parallelFuc
A process that will be run in parallel

listitems
The main list for each process

unprocessedListitems
The remain list items that haven't been processed yet

...
other parameters

Value

No return

Examples

coming later

[Package FBNNet version 0.0.0.9000 Index]