extractGeneStateFromTimeSeriesCube {FBNNet}	R Documentation

An internal method to extract gene state from timeseries cube

Description

An internal method to extract gene state from timeseries cube

Usage

extractGeneStateFromTimeSeriesCube(timeSeriesCube)

Arguments

timeSeriesCube
A list object contains samples in which a sample contains genes in row and time points in column

Value

A combined matrix

Examples

mat1<-matrix(c("1","0","0","1","0","0","0","1","1"),3,3, dimnames=list(c("gene1","gene2","gene3"),c("1","2","3")))
mat2<-matrix(c("1","1","0","1","0","1","1","1","0"),3,3, dimnames=list(c("gene1","gene2","gene3"),c("1","2","3")))
listtest<-list(mat1,mat2)
extractGeneStateFromTimeSeriesCube(listtest)

[Package FBNNet version 0.0.0.9000 Index]