getFBMSuccessor {FBNNet}	R Documentation

This method is used to calculate the next state

Description

This method is used to calculate the next state

Usage

getFBMSuccessor(fbnNetwork, currentState, genes, type = c("synchronous",
  "asynchronous"), decayIndex = c(), timestepTrack = list())

Arguments

currentState
A vector of current gene state

genes
a list of genes which index order must match with the current state

type
A type of Boolean network update schema choosen from synchronous, asynchronous and time step based

timestepTrack
A list of processed time steps for each FBM function. This parameter is used with the type of timestep

fbnnetwork
An object of FBNNetwork

timedecay
An value indicates the period of time when to degrade an activated gene if no activators presented. It is usually one time step

Value

A list object that contains reconstructed time series and FBN network

Examples

require(BoolNet)
require(FBNNet)
network<-loadNetwork("testthat/others/cellcycle.txt")
trainingseries<-FBNDataReduction(generateTimeSeries(network,2000,43))
cube<-constructFBNCube(network$genes,network$genes,trainingseries,4,1,TRUE)
NETWORK2<-mineFBNNetwork(cube,network$genes)
state<-c("0","1","1","0","1","1","1","1","0","0")
names(state)<-c("CycD","Rb","E2F","CycE","CycA","p27","Cdc20","Cdh1","UbcH10","CycB")
getFBMSuccessor(NETWORK2,state,names(state))

[Package FBNNet version 0.0.0.9000 Index]