getGenePrababilities {FBNNet}	R Documentation

A core method to calculate all measures for a target gene with a condition gene

Description

A core method to calculate all measures for a target gene with a condition gene

Usage

getGenePrababilities(timeSeriesCube, fixedgenestate, targetgene, newgene,
  geneStates = NULL, preGeneStates = NULL, temporal = 1,
  temporalGeneStates = NULL)

Arguments

timeSeriesCube
A list object contains samples in which a sample contains genes in row and time points in column

fixedgenestate
prefixed gene state

targetgene
The tage gene's name

newgene
The new added conditional gene

geneStates
The current combined gene state

preGeneStates
The previous combined gene state

temporal
The temporal setting and the default value is 1

temporalGeneStates
The combined gene state for temporal

Value

A list of measures

Examples

genesInput<-c("CycD","p27","CycE","E2F")
testseries<-list()
testseries[[1]]<-matrix(c(1,1,1,1,0,1,0,1,1,1,1,1,0,0,0,1),nrow=4,ncol=4,byrow=FALSE,dimnames<-list(genesInput,c("1","2","3","4")))
testseries[[2]]<-matrix(c(1,1,0,1,0,1,0,1,0,1,0,1,1,0,0,1),nrow=4,ncol=4,byrow=FALSE,dimnames<-list(genesInput,c("1","2","3","4")))
testseries[[3]]<-matrix(c(1,0,0,1,1,1,0,1,1,1,0,1,1,1,0,1),nrow=4,ncol=4,byrow=FALSE,dimnames<-list(genesInput,c("1","2","3","4")))

getCurrentState<-extractGeneStateFromTimeSeriesCube(testseries)
getCurrentState<-getCurrentState[,-1]#remove the first one

getpreviousState<-extractGeneStateFromTimeSeriesCube(testseries)
getpreviousState<-getpreviousState[,-length(colnames(getpreviousState))] #remove the last one
probability<-getGenePrababilities(testseries,NULL,"CycD","p27",getCurrentState,getpreviousState,1)

[Package FBNNet version 0.0.0.9000 Index]