getProbabilityFromCubeBasedOnInput {FBNNet}	R Documentation

An internal method to get the probability of regulatory function from the cube

Description

An internal method to get the probability of regulatory function from the cube

Usage

getProbabilityFromCubeBasedOnInput(targetgene, funcType, FBNExpression,
  FBNProbability, preGeneInputs)

Arguments

targetgene
The target gene

funcType
type of function

FBNExpression
expression of function

preGeneInputs
pre input genes' state

FBNCube
An object of orchard cube

Value

A probablity of the target regulatory function

Examples

coming later

[Package FBNNet version 0.0.0.9000 Index]