getncbiseq {FBNNet}	R Documentation

Retrieving genome sequence data using SeqinR. The code is a sample from the book "a little book of r for bioinformatic"

Description

Retrieving genome sequence data using SeqinR. The code is a sample from the book "a little book of r for bioinformatic"

Usage

getncbiseq(accession)

Arguments

accession
NBCI accession number

Value

genome sequence

Examples

coming later

[Package FBNNet version 0.0.0.9000 Index]