identifyDifferentiallyExpressedGenes {FBNNet}	R Documentation

A benchmark type function to test a complete process of FBN model Step 3, identify significantly expressed genes, which are strongly related with the samples / study purposes

Description

A benchmark type function to test a complete process of FBN model Step 3, identify significantly expressed genes, which are strongly related with the samples / study purposes

Usage

identifyDifferentiallyExpressedGenes(orderSampleTimeSeries)

[Package FBNNet version 0.0.0.9000 Index]