reconstructTimeseries {FBNNet}	R Documentation

This method is used to reconstruct time series data

Description

This method is used to reconstruct time series data

Usage

reconstructTimeseries(fbnCube, fbnnetwork, timeseriesCube, interval = 1,
  maxTimepoints = 100)

Arguments

fbnCube
A pre constructed Orchard cube

fbnnetwork
An object of FBNNetwork

timeseriesCube
An original time series data that contains samples, genes and time points. It can be as list of samples contain genes and the initial time point.

interval
An interval is the time gap between two time points

Value

A list object that contains reconstructed time series and FBN network

Examples

mat1<-matrix(c("1","0","0","1","0","0","0","1","1"),3,3, dimnames=list(c("gene1","gene2","gene3"),c("1","2","3")))
mat2<-matrix(c("1","1","0","1","0","1","1","1","0"),3,3, dimnames=list(c("gene1","gene2","gene3"),c("1","2","3")))
listtestOriginal<-list(mat1,mat2)
cube<-constructFBNCube(c("gene1","gene2"),c("gene1","gene2","gene3"),listtestOriginal,4,1,FALSE)
network<-mineFBNNetwork(cube,c("gene1","gene2"))
mat1<-matrix(c("1","1","0"),3,1, dimnames=list(c("gene1","gene2","gene3"),c("1")))
mat2<-matrix(c("1","1","1"),3,1, dimnames=list(c("gene1","gene2","gene3"),c("1")))
listtestInitial<-list(mat1,mat2)
result<-reconstructTimeseries(cube,network,listtestInitial,1,1,3)

[Package FBNNet version 0.0.0.9000 Index]