timeseriesReduction {FBNNet}	R Documentation

Remove some time points out from the target time series

Description

Remove some time points out from the target time series

Usage

timeseriesReduction(timeseriesCube, reducedTimePoints)

Arguments

timeseriesCube
A data set that contains samples and each sample contains genes and time points

reducedTimePoints
How many time points reduced between two observation time points

Value

reduced time series cube

Examples

coming later

[Package FBNNet version 0.0.0.9000 Index]