Executing a list of processes in parallel with decreasing list

Executing a list of processes in parallel with decreasing list

FBNNetwork.Graph(
  fbnNetwork,
  type = "static",
  timeseriesMatrix = NULL,
  fromTimePoint = 1,
  toTimePoint = 5,
  networkobject = NULL
)

Arguments

fbnNetwork

A FBN Network object

type

A type of object and its value is type=static | dynamic | staticSlice

timeseriesMatrix

A sample timeseries matrix

fromTimePoint

The time point at the beginning

toTimePoint

The time point at the end

networkobject

An object created by the funciton ConvertToNetworkGraphicObject.

Value

No return

Examples

data('yeastTimeSeries')
network <- generateFBMNetwork(yeastTimeSeries)
network
#> Fundamental Boolean Network with  3 genes
#> Genes involved:
#> Fkh2, Swi5, Clb1
#> 
#> Networks:
#> Multiple Transition Functions for Fkh2 with decay value = 1:
#> Fkh2_1_Activator: Fkh2 = Fkh2 (Confidence: 1, TimeStep: 1)
#> Fkh2_2_Activator: Fkh2 = Clb1 (Confidence: 1, TimeStep: 1)
#> 
#> Multiple Transition Functions for Swi5 with decay value = 1:
#> Swi5_1_Activator: Swi5 = Fkh2 (Confidence: 1, TimeStep: 1)
#> Swi5_2_Activator: Swi5 = Clb1 (Confidence: 1, TimeStep: 1)
#> 
#> Multiple Transition Functions for Clb1 with decay value = 1:
#> Clb1_1_Activator: Clb1 = Fkh2 (Confidence: 1, TimeStep: 1)
#> Clb1_2_Activator: Clb1 = Clb1 (Confidence: 1, TimeStep: 1)
## draw the general graph
FBNNetwork.Graph(network)