Create an Orchard cube

This is the main function(s) to genereate a single Orchard Cube or a group of cubes

constructFBNCube(
  target_genes,
  conditional_genes,
  timeseriesCube,
  maxK = 5,
  temporal = 1,
  useParallel = FALSE
)

Arguments

target_genes

A vector of genes that will be treated as target genes

conditional_genes

All genes that are available for building up the cube

timeseriesCube

A list of samples in which a sample is a matrix that contains gene states where genes in rows and time points in columns.

maxK

The maximum level the cube can dig in

temporal

A value that used to be 1 indicates the previous steps the current one can depend on

useParallel

If it is TRUE, the constructing will run it in parallel, otherwise in a singl thread

Value

An Orchard cube that contains all precomputed measures

Examples

require(BoolNet)
#> Loading required package: BoolNet
data('ExampleNetwork')
initialStates <- generateAllCombinationBinary(ExampleNetwork$genes)
trainingseries <- genereateBoolNetTimeseries(ExampleNetwork,
                                           initialStates,
                                           43,
                                           type='synchronous')
cube<-constructFBNCube(target_genes = ExampleNetwork$genes,
                       conditional_genes = ExampleNetwork$genes,
                       timeseriesCube = trainingseries,
                       maxK = 4,
                       temporal = 1,
                       useParallel = FALSE)
#> INFO [2022-04-08 11:04:33] Enter constructFBNCube zone: 
#>         target_genes=5 genes and they are Gene1, Gene2, Gene3, Gene4, Gene5,
#>         conditional_genes=5 genes and they are Gene1, Gene2, Gene3, Gene4, Gene5,
#>         data_length=32,
#>         maxK=4, 
#>         temporal=1,
#>         useParallel=FALSE
#> INFO [2022-04-08 11:04:33] Leave constructFBNCube zone.
network <- mineFBNNetwork(cube,ExampleNetwork$genes)
#> INFO [2022-04-08 11:04:33] Enter mineFBNNetwork zone: genes=Gene1, Gene2, Gene3, Gene4, Gene5, useParallel=FALSE
#> INFO [2022-04-08 11:04:33] Enter search_FBN_core zone: useParallel=FALSE
#> INFO [2022-04-08 11:04:33] Leave search_FBN_core zone
#> INFO [2022-04-08 11:04:33] Enter mineFBNNetworkWithCores zone
#> INFO [2022-04-08 11:04:33] Enter mineFBNNetworkStage2 zone
#> INFO [2022-04-08 11:04:33] Leave mineFBNNetworkStage2 zone
#> INFO [2022-04-08 11:04:33] Enter convertMinedResultToFBNNetwork zone
#> INFO [2022-04-08 11:04:33] Leave convertMinedResultToFBNNetwork zone
#> INFO [2022-04-08 11:04:33] Leave mineFBNNetworkWithCores zone
#> INFO [2022-04-08 11:04:33] Leave mineFBNNetwork zone
network
#> Fundamental Boolean Network with  5 genes
#> Genes involved:
#> Gene1, Gene2, Gene3, Gene4, Gene5
#> 
#> Networks:
#> Multiple Transition Functions for Gene1 with decay value = 1:
#> Gene1_1_Activator: Gene1 = Gene1 (Confidence: 1, TimeStep: 1)
#> Gene1_1_Inhibitor: Gene1 = !Gene1 (Confidence: 1, TimeStep: 1)
#> 
#> Multiple Transition Functions for Gene2 with decay value = 1:
#> Gene2_1_Activator: Gene2 = Gene1&!Gene4&Gene5 (Confidence: 1, TimeStep: 1)
#> Gene2_1_Inhibitor: Gene2 = !Gene1 (Confidence: 1, TimeStep: 1)
#> Gene2_2_Inhibitor: Gene2 = Gene4 (Confidence: 1, TimeStep: 1)
#> Gene2_3_Inhibitor: Gene2 = !Gene5 (Confidence: 1, TimeStep: 1)
#> 
#> Multiple Transition Functions for Gene3 with decay value = 1:
#> Gene3_1_Activator: Gene3 = Gene3 (Confidence: 1, TimeStep: 1)
#> Gene3_1_Inhibitor: Gene3 = !Gene3 (Confidence: 1, TimeStep: 1)
#> 
#> Multiple Transition Functions for Gene4 with decay value = 1:
#> Gene4_1_Activator: Gene4 = !Gene1&Gene3 (Confidence: 1, TimeStep: 1)
#> Gene4_2_Activator: Gene4 = Gene3&!Gene5 (Confidence: 1, TimeStep: 1)
#> Gene4_1_Inhibitor: Gene4 = !Gene3 (Confidence: 1, TimeStep: 1)
#> Gene4_2_Inhibitor: Gene4 = Gene1&Gene5 (Confidence: 1, TimeStep: 1)
#> 
#> Multiple Transition Functions for Gene5 with decay value = 1:
#> Gene5_1_Activator: Gene5 = !Gene2 (Confidence: 1, TimeStep: 1)
#> Gene5_1_Inhibitor: Gene5 = Gene2 (Confidence: 1, TimeStep: 1)
## draw the general graph
FBNNetwork.Graph(network)