This method is used to calculate the next state

This method is used to calculate the next state

getFBMSuccessor(
  fbnNetwork,
  previous_states,
  current_step,
  genes,
  type = c("synchronous", "asynchronous"),
  decayIndex = c()
)

Arguments

fbnNetwork
An object of FBNNetwork

previous_states
A vector of current gene state

current_step
The index of the current step

genes
a list of genes which index order must match with the current state

type
A type of Boolean network update schema chosen from synchronous, asynchronous based. Asynchronous will randomly pick up a gene to process at time.

decayIndex
An value indicates the period of time when to degrade an activated gene if no activators presented. It is usually one time step

Value

A list object that contains reconstructed time series and FBN network

Examples

require(BoolNet)
data(ExampleNetwork)
trainingseries<-FBNDataReduction(generateTimeSeries(ExampleNetwork,2000,43))
cube<-constructFBNCube(target_genes = ExampleNetwork$genes,
                       conditional_genes = ExampleNetwork$genes,
                       timeseriesCube = trainingseries,
                       maxK = 3,
                       temporal = 1,
                       useParallel = FALSE)
NETWORK2<-mineFBNNetwork(cube,ExampleNetwork$genes)
state<-c('0','1','1','0','1')
names(state)<-c('Gene1','Gene2','Gene3','Gene4','Gene5')
getFBMSuccessor(NETWORK2,
                previous_states= state,
                current_step = 2,
                genes = names(state),
                type = 'synchronous')

#> $nextState
#> Gene1 Gene2 Gene3 Gene4 Gene5 
#>     0     0     1     1     0 
#> 
#> $decayIndex
#> Gene1 Gene2 Gene3 Gene4 Gene5 
#>     1     1     1     1     1 
#>