This method is used to reconstruct time series data

This method is used to reconstruct time series data

reconstructTimeseries(
  fbnnetwork,
  initialStates,
  type = c("synchronous", "asynchronous"),
  maxTimepoints = 100,
  useParallel = FALSE
)

Arguments

fbnnetwork
An object of FBN network

initialStates
A list of initial states

type
Specify the type of the Fundamental Boolean model (synchronous or asynchronous)

maxTimepoints
The max time points that are going to be constructed

useParallel
Optional, if TRUE then use parallelisation

Value

A list object that contains reconstructed time series and FBN network

Examples

require(BoolNet)
data('ExampleNetwork')
initialStates <- generateAllCombinationBinary(ExampleNetwork$genes)
trainingseries <- genereateBoolNetTimeseries(ExampleNetwork,
                                           initialStates,
                                           43,
                                           type='synchronous')
cube <- constructFBNCube(target_genes = ExampleNetwork$genes,
                       conditional_genes = ExampleNetwork$genes,
                       timeseriesCube = trainingseries,
                       maxK = 4,
                       temporal = 1,
                       useParallel = FALSE)
network <- mineFBNNetwork(cube)
result <- reconstructTimeseries(fbnnetwork = network,
                              initialStates = initialStates,
                              type = 'synchronous',
                              maxTimepoints = 43)
result

#> [[1]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5 0 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> 
#> [[2]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2 0 0 1 1 0 0 1 1 0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1
#> Gene3 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5 0 1 1 0 0 1 1 0 0  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1
#> Gene3  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1
#> 
#> [[3]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2 1 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5 0 0 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> 
#> [[4]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2 1 0 0 1 1 0 0 1 1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0
#> Gene3 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5 0 0 1 1 0 0 1 1 0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0
#> Gene3  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1
#> 
#> [[5]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4 0 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene5 0 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene5  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> 
#> [[6]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2 0 0 0 1 1 0 0 0 1  1  0  0  0  1  1  0  0  0  1  1  0  0  0  1  1  0  0
#> Gene3 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4 0 1 0 0 0 1 1 0 0  0  1  1  0  0  0  1  1  0  0  0  1  1  0  0  0  1  1
#> Gene5 0 1 1 1 0 0 1 1 1  0  0  1  1  1  0  0  1  1  1  0  0  1  1  1  0  0  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2  0  1  1  0  0  0  1  1  0  0  0  1  1  0  0  0
#> Gene3  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4  0  0  0  1  1  0  0  0  1  1  0  0  0  1  1  0
#> Gene5  1  1  0  0  1  1  1  0  0  1  1  1  0  0  1  1
#> 
#> [[7]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2 1 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4 0 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene5 0 0 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene5  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> 
#> [[8]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2 1 0 0 0 1 1 0 0 0  1  1  0  0  0  1  1  0  0  0  1  1  0  0  0  1  1  0
#> Gene3 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4 0 1 1 0 0 0 1 1 0  0  0  1  1  0  0  0  1  1  0  0  0  1  1  0  0  0  1
#> Gene5 0 0 1 1 1 0 0 1 1  1  0  0  1  1  1  0  0  1  1  1  0  0  1  1  1  0  0
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2  0  0  1  1  0  0  0  1  1  0  0  0  1  1  0  0
#> Gene3  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4  1  0  0  0  1  1  0  0  0  1  1  0  0  0  1  1
#> Gene5  1  1  1  0  0  1  1  1  0  0  1  1  1  0  0  1
#> 
#> [[9]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4 1 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5 0 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> 
#> [[10]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2 0 0 1 1 0 0 1 1 0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1
#> Gene3 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4 1 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5 0 1 1 0 0 1 1 0 0  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1
#> Gene3  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1
#> 
#> [[11]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2 1 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4 1 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5 0 0 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> 
#> [[12]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2 1 0 0 1 1 0 0 1 1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0
#> Gene3 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4 1 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5 0 0 1 1 0 0 1 1 0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0
#> Gene3  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1
#> 
#> [[13]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene5 0 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene5  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> 
#> [[14]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2 0 0 0 1 1 0 0 0 1  1  0  0  0  1  1  0  0  0  1  1  0  0  0  1  1  0  0
#> Gene3 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4 1 1 0 0 0 1 1 0 0  0  1  1  0  0  0  1  1  0  0  0  1  1  0  0  0  1  1
#> Gene5 0 1 1 1 0 0 1 1 1  0  0  1  1  1  0  0  1  1  1  0  0  1  1  1  0  0  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2  0  1  1  0  0  0  1  1  0  0  0  1  1  0  0  0
#> Gene3  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4  0  0  0  1  1  0  0  0  1  1  0  0  0  1  1  0
#> Gene5  1  1  0  0  1  1  1  0  0  1  1  1  0  0  1  1
#> 
#> [[15]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2 1 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene5 0 0 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene5  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> 
#> [[16]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2 1 0 0 0 1 1 0 0 0  1  1  0  0  0  1  1  0  0  0  1  1  0  0  0  1  1  0
#> Gene3 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4 1 1 1 0 0 0 1 1 0  0  0  1  1  0  0  0  1  1  0  0  0  1  1  0  0  0  1
#> Gene5 0 0 1 1 1 0 0 1 1  1  0  0  1  1  1  0  0  1  1  1  0  0  1  1  1  0  0
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2  0  0  1  1  0  0  0  1  1  0  0  0  1  1  0  0
#> Gene3  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4  1  0  0  0  1  1  0  0  0  1  1  0  0  0  1  1
#> Gene5  1  1  1  0  0  1  1  1  0  0  1  1  1  0  0  1
#> 
#> [[17]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> 
#> [[18]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2 0 1 1 0 0 1 1 0 0  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1
#> Gene3 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5 1 1 0 0 1 1 0 0 1  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1
#> Gene3  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0
#> 
#> [[19]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2 1 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5 1 0 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> 
#> [[20]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2 1 1 0 0 1 1 0 0 1  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0
#> Gene3 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5 1 0 0 1 1 0 0 1 1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0
#> Gene3  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0
#> 
#> [[21]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4 0 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene5 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene5  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> 
#> [[22]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2 0 1 1 0 0 0 1 1 0  0  0  1  1  0  0  0  1  1  0  0  0  1  1  0  0  0  1
#> Gene3 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4 0 0 0 1 1 0 0 0 1  1  0  0  0  1  1  0  0  0  1  1  0  0  0  1  1  0  0
#> Gene5 1 1 0 0 1 1 1 0 0  1  1  1  0  0  1  1  1  0  0  1  1  1  0  0  1  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2  1  0  0  0  1  1  0  0  0  1  1  0  0  0  1  1
#> Gene3  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4  0  1  1  0  0  0  1  1  0  0  0  1  1  0  0  0
#> Gene5  0  0  1  1  1  0  0  1  1  1  0  0  1  1  1  0
#> 
#> [[23]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2 1 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4 0 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene5 1 0 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene5  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> 
#> [[24]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2 1 1 0 0 0 1 1 0 0  0  1  1  0  0  0  1  1  0  0  0  1  1  0  0  0  1  1
#> Gene3 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4 0 0 1 1 0 0 0 1 1  0  0  0  1  1  0  0  0  1  1  0  0  0  1  1  0  0  0
#> Gene5 1 0 0 1 1 1 0 0 1  1  1  0  0  1  1  1  0  0  1  1  1  0  0  1  1  1  0
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2  0  0  0  1  1  0  0  0  1  1  0  0  0  1  1  0
#> Gene3  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4  1  1  0  0  0  1  1  0  0  0  1  1  0  0  0  1
#> Gene5  0  1  1  1  0  0  1  1  1  0  0  1  1  1  0  0
#> 
#> [[25]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4 1 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> 
#> [[26]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2 0 0 1 1 0 0 1 1 0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1
#> Gene3 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4 1 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5 1 1 1 0 0 1 1 0 0  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1
#> Gene3  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1
#> 
#> [[27]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2 1 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4 1 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5 1 0 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> 
#> [[28]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2 1 0 0 1 1 0 0 1 1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0
#> Gene3 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4 1 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5 1 0 1 1 0 0 1 1 0  0  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2  1  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0
#> Gene3  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene4  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene5  1  0  0  1  1  0  0  1  1  0  0  1  1  0  0  1
#> 
#> [[29]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene5 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene5  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> 
#> [[30]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2 0 0 1 1 0 0 0 1 1  0  0  0  1  1  0  0  0  1  1  0  0  0  1  1  0  0  0
#> Gene3 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4 1 0 0 0 1 1 0 0 0  1  1  0  0  0  1  1  0  0  0  1  1  0  0  0  1  1  0
#> Gene5 1 1 1 0 0 1 1 1 0  0  1  1  1  0  0  1  1  1  0  0  1  1  1  0  0  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2  1  1  0  0  0  1  1  0  0  0  1  1  0  0  0  1
#> Gene3  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4  0  0  1  1  0  0  0  1  1  0  0  0  1  1  0  0
#> Gene5  1  0  0  1  1  1  0  0  1  1  1  0  0  1  1  1
#> 
#> [[31]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 0 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2 1 0 0 0 0 0 0 0 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene5 1 0 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene2  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0
#> Gene3  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene5  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> 
#> [[32]]
#>       1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27
#> Gene1 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2 1 0 0 0 1 1 0 0 0  1  1  0  0  0  1  1  0  0  0  1  1  0  0  0  1  1  0
#> Gene3 1 1 1 1 1 1 1 1 1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4 1 0 1 0 0 0 1 1 0  0  0  1  1  0  0  0  1  1  0  0  0  1  1  0  0  0  1
#> Gene5 1 0 1 1 1 0 0 1 1  1  0  0  1  1  1  0  0  1  1  1  0  0  1  1  1  0  0
#>       28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43
#> Gene1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene2  0  0  1  1  0  0  0  1  1  0  0  0  1  1  0  0
#> Gene3  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> Gene4  1  0  0  0  1  1  0  0  0  1  1  0  0  0  1  1
#> Gene5  1  1  1  0  0  1  1  1  0  0  1  1  1  0  0  1
#> 
#> attr(,"class")
#> [1] "FBNTimeSeries"